requirements.txt
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
rich>=13.7.0
composio-core>=0.3.0
requests>=2.31.0

